Alternative approach based on knitr::kable and kableExtra.
This allows to have better control on the formatting of the tables for both
html and pdf outputs while avoiding duplicate code for the two formats.

This also removes the few other uses of Latex \tabular, so that more missing
tables will now appear in the html document.

Since Latex minipages are not displayed in the html version if used directly,
in sparse-ragged.Rmd we use knitr:::pandoc_to() to check the format being
generated: if html, then we issue standard html table markup, otherwise we
use \minipage.

The two uses of \caption have been removed and the text now flows as a
normal paragraph (idea taken from @enbrown's PR).